Fix round() on double precision in the matcher trigger
notify_nearby_providers built its notification body with
round(ST_Distance(...) / 1000.0, 1). ST_Distance returns double precision,
and Postgres has no round(double precision, integer) -- only the numeric
form. Every job insert therefore failed with:

  function round(double precision, integer) does not exist

Cast the division to numeric before rounding. The single-argument
round(double precision) on the next line is a different function and was
always fine, as were the round(numeric, 2) calls in complete_job.

Found by applying the migrations to a real Postgres for the first time;
the full job lifecycle passes after this fix.

+++ supabase/migrations/20260922090200_functions_and_triggers.sql
@@ -168,7 +168,8 @@ begin
     'new_job_nearby',
     case when new.is_urgent then 'Urgent ' || v_cat_en || ' job nearby'
          else 'New ' || v_cat_en || ' job nearby' end,
-    new.title || ' - ' || round(extensions.st_distance(pp.current_location, new.location) / 1000.0, 1) || ' km away',
+    -- ST_Distance returns double precision; round(double, int) does not exist.
+    new.title || ' - ' || round((extensions.st_distance(pp.current_location, new.location) / 1000.0)::numeric, 1) || ' km away',
     new.id,
     jsonb_build_object(
       'distance_m', round(extensions.st_distance(pp.current_location, new.location)),
